Replaces `tp.Compiler` with `tp.compile`
Removes the `tp.Compiler` class and instead replaces it with just a standalone
function, `tp.compile`.

Also splits up the `backend/api` tests to reflect the structure of the code.
