+++ tripy/docs/pre0_user_guides/02-compiler.md
@@ -28,17 +28,9 @@ inp = tp.ones((1, 2))
 out = layer(inp)
 ```
 
-Now, let's try to optimize this model for inference using Tripy's {class}`tripy.Compiler`.
+Now, let's try to optimize this model for inference using Tripy's {func}`tripy.compile`.
 
-First, let's initialize the compiler with the module we want to compile, `layer`,
-which lets the compiler know its properties, like the function signature.
-
-```py
-# doc: no-print-locals
-compiler = tp.Compiler(layer)
-```
-
-Next, we need to provide information about each input using {class}`tripy.InputInfo`.
+When we compile our module, we need to provide information about each input using {class}`tripy.InputInfo`.
 The first argument for `InputInfo` is `shape`, where we specify either the static or
 dynamic shape information for each dimension. In the example below, we assume the
 shape of `inp` is static (`(1, 2)`). The second argument specifies the `dtype` for the input:
@@ -51,7 +43,7 @@ Now, we can call the `compile` function to obtain a compiled function and use it
 
 ```py
 # doc: no-print-locals
-fast_geglu = compiler.compile(inp_info)
+fast_geglu = tp.compile(layer, args=[inp_info])
 fast_geglu(inp).eval()
 ```
 
@@ -67,7 +59,7 @@ and it should optimize for a size of 8.
 ```py
 # doc: print-locals out out_change_shape
 inp_info = tp.InputInfo(shape=((1, 8, 16), 2), dtype=tp.float32)
-fast_geglu = compiler.compile(inp_info)
+fast_geglu = tp.compile(layer, args=[inp_info])
 out = fast_geglu(inp)
 
 # Let's change the shape of input to (2, 2)
@@ -94,9 +86,11 @@ Saving an executable to disk:
 
 ```py
 # doc: no-print-locals
-import tempfile, os
-temp_dir = tempfile.mkdtemp()
-executable_file_path = os.path.join(temp_dir, "executable.json")
+import tempfile # doc: omit
+import os
+
+out_dir = tempfile.mkdtemp() # doc: omit
+executable_file_path = os.path.join(out_dir, "executable.json")
 fast_geglu.save(executable_file_path)
 ```
 
@@ -104,10 +98,11 @@ Reading an executable and running inference:
 
 ```py
 # doc: no-print-locals
-inp = tp.Tensor([[1., 2.], [2., 3.]], dtype=tp.float32)
 loaded_fast_geglu = tp.Executable.load(executable_file_path)
+
+inp = tp.Tensor([[1., 2.], [2., 3.]], dtype=tp.float32)
 out = loaded_fast_geglu(inp)
-os.remove(executable_file_path)
+os.remove(executable_file_path) # doc: omit
 ```
 
 ### Querying Executable Properties
@@ -134,9 +129,8 @@ def add_times_two(a, b):
     print(f"c : {c}")
     return c + a + b
 
-compiler = tp.Compiler(add_times_two)
 inp_info = tp.InputInfo(shape=(1, 2), dtype=tp.float32)
-fast_myadd = compiler.compile(inp_info, inp_info)
+fast_myadd = tp.compile(add_times_two, args=[inp_info, inp_info])
 a = tp.Tensor([[1.0, 2.0]], dtype=tp.float32)
 b = tp.Tensor([[2.0, 3.0]], dtype=tp.float32)
