+++ tripy/tripy/backend/api/executable.py
@@ -28,12 +28,12 @@ from tripy.frontend import Tensor
 from tripy.utils import json as json_utils
 
 
-@export.public_api(document_under="compiler")
+@export.public_api(document_under="compiling_code")
 class Executable:
     """
     Represents a compiled executable generated by the compiler.
 
-    .. seealso:: :class:`Compiler`
+    .. seealso:: :func:`compile`
     """
 
     # The constructor is intentionally undocumented because it is not meant to be called by users.
@@ -83,15 +83,16 @@ class Executable:
             def add(a, b):
                 return a + b
 
-            # doc: no-print-locals compiler compiled_add
-            compiler = tp.Compiler(add)
-            compiled_add = compiler.compile(tp.InputInfo((1,), dtype=tp.float32), tp.InputInfo((1,), dtype=tp.float32))
+            # doc: no-print-locals compiled_add
+            compiled_add = tp.compile(add, args=[tp.InputInfo((1,), dtype=tp.float32), tp.InputInfo((1,), dtype=tp.float32)])
 
             a = tp.ones((1,), dtype=tp.float32)
             b = tp.ones((1,), dtype=tp.float32)
 
             out = compiled_add(a, b)
         """
+        NUM_ARGS = len(args) + len(kwargs)
+
         input_tensors = []
 
         input_tensors.extend(args)
@@ -101,7 +102,7 @@ class Executable:
             if name not in kwargs:
                 raise_error(f"Missing argument: {name}", [f"Expected the following arguments: {self._arg_names}"])
 
-            input_tensors.extend(kwargs[name])
+            input_tensors.append(kwargs[name])
             del kwargs[name]
 
         if kwargs:
@@ -114,13 +115,11 @@ class Executable:
             )
 
         # We do this after kwarg checks since those will be more informative (we can explain which arguments are missing/extra).
-        num_args = len(args) + len(kwargs)
-
-        if num_args != len(self._arg_names):
+        if NUM_ARGS != len(self._arg_names):
             raise_error(
                 "Incorrect number of arguments.",
                 [
-                    f"Expected {len(self._arg_names)} arguments but got {num_args}.\n"
+                    f"Expected {len(self._arg_names)} arguments but got {NUM_ARGS}.\n"
                     f"Note: Expected arguments were: {self._arg_names}",
                 ],
             )
@@ -194,9 +193,8 @@ class Executable:
             def add(a, b):
                 return a + b
 
-            # doc: no-print-locals compiler compiled_add
-            compiler = tp.Compiler(add)
-            compiled_add = compiler.compile(tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32))
+            # doc: no-print-locals compiled_add
+            compiled_add = tp.compile(add, args=[tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32)])
             print(compiled_add.get_input_info())
         """
         input_info = []
@@ -218,9 +216,8 @@ class Executable:
             def add(a, b):
                 return a + b
 
-            # doc: no-print-locals compiler compiled_add
-            compiler = tp.Compiler(add)
-            compiled_add = compiler.compile(tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32))
+            # doc: no-print-locals compiled_add
+            compiled_add = tp.compile(add, args=[tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32)])
             print(compiled_add.get_output_info())
         """
         output_info = []
@@ -245,9 +242,8 @@ class Executable:
             def add(a, b):
                 return a + b
 
-            # doc: no-print-locals compiler compiled_add executable_file
-            compiler = tp.Compiler(add)
-            compiled_add = compiler.compile(tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32))
+            # doc: no-print-locals compiled_add executable_file
+            compiled_add = tp.compile(add, args=[tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32)])
 
             with tempfile.TemporaryDirectory() as temp_dir:
                 executable_file = os.path.join(temp_dir, "executable.json")
@@ -276,9 +272,8 @@ class Executable:
             def add(a, b):
                 return a + b
 
-            # doc: no-print-locals compiler compiled_add executable_file
-            compiler = tp.Compiler(add)
-            compiled_add = compiler.compile(tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32))
+            # doc: no-print-locals compiled_add executable_file
+            compiled_add = tp.compile(add, args=[tp.InputInfo(([1, 2, 3],), dtype=tp.float32), tp.InputInfo(([1, 2, 3],), dtype=tp.float32)])
 
             with tempfile.TemporaryDirectory() as temp_dir:
                 executable_file = os.path.join(temp_dir, "executable.json")
