+++ tripy/tripy/backend/api/input_info.py
@@ -22,7 +22,7 @@ from tripy.common.exception import raise_error
 from tripy.common.shape_bounds import ShapeBounds
 
 
-@export.public_api(document_under="compiler")
+@export.public_api(document_under="compiling_code")
 class InputInfo:
     """
     Captures information about an input to a compiled function.
@@ -95,7 +95,7 @@ class InputInfo:
 
 
 # TODO(MLIR-TRT #923): Can generalize `InputInfo` and drop this class.
-@export.public_api(document_under="compiler")
+@export.public_api(document_under="compiling_code")
 @dataclass
 class ArgInfo:
     shape_bounds: Sequence[Tuple[int, int]]
